refactor(any2mochi): move java converter
