+++ tools/any2mochi/x/java/convert.go
@@ -1,4 +1,4 @@
-package any2mochi
+package java
 
 import (
 	"fmt"
@@ -6,12 +6,12 @@ import (
 	"strings"
 )
 
-// ConvertJava converts a very small subset of Java source code to Mochi.
+// Convert converts a very small subset of Java source code to Mochi.
 // The converter is intentionally limited but is able to handle the simple
 // Java programs generated by the Mochi compiler. Unsupported constructs
 // result in an informative error.
-func ConvertJava(src string) ([]byte, error) {
-	lines, err := parseJava(src)
+func Convert(src string) ([]byte, error) {
+	lines, err := parse(src)
 	if err != nil {
 		return nil, err
 	}
@@ -306,11 +306,22 @@ func parseFieldLine(line string) (string, string) {
 	return strings.TrimSpace(line[:idx]), strings.TrimSpace(line[idx+1:])
 }
 
-// ConvertJavaFile reads the java file and converts it to Mochi.
-func ConvertJavaFile(path string) ([]byte, error) {
+// ConvertFile reads the Java file and converts it to Mochi.
+func ConvertFile(path string) ([]byte, error) {
 	data, err := os.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
-	return ConvertJava(string(data))
+	return Convert(string(data))
+}
+
+func numberedSnippet(src string) string {
+	lines := strings.Split(src, "\n")
+	if len(lines) > 10 {
+		lines = lines[:10]
+	}
+	for i, l := range lines {
+		lines[i] = fmt.Sprintf("%3d: %s", i+1, l)
+	}
+	return strings.Join(lines, "\n")
 }
